Move testmatrix.h in the remaining tests.
This commit finishes the work started in bea2270a943, where
tests/lac/testmatrix.h was renamed to tests/testmatrix.h.
