+++ tests/lapack/solver_cg.cc
@@ -21,7 +21,7 @@
 #include <fstream>
 #include <iostream>
 #include <iomanip>
-#include "../lac/testmatrix.h"
+#include "../testmatrix.h"
 #include <deal.II/base/logstream.h>
 #include <deal.II/lac/sparse_matrix.h>
 #include <deal.II/lac/sparse_matrix.templates.h>
