+++ tests/petsc/solver_03_precondition_boomeramg_symmetric.cc
@@ -20,7 +20,7 @@
 
 
 #include "../tests.h"
-#include "../lac/testmatrix.h"
+#include "../testmatrix.h"
 #include <cmath>
 #include <fstream>
 #include <iostream>
