+++ tests/petsc/solver_03_precondition_sor.cc
@@ -18,7 +18,7 @@
 
 
 #include "../tests.h"
-#include "../lac/testmatrix.h"
+#include "../testmatrix.h"
 #include <cmath>
 #include <fstream>
 #include <iostream>
