+++ tests/slepc/solve_04.cc
@@ -20,7 +20,7 @@
 
 
 #include "../tests.h"
-#include "../lac/testmatrix.h"
+#include "../testmatrix.h"
 #include "testmatrix.h"
 #include <cmath>
 #include <fstream>
